Use `setup-python` in coverage report job in CI workflow
- Remove the container declaration from the `coverage_report` job.
- Add a `setup-python` step to `coverage_report` for Python
  installation.
- Define `PYTHON_VERSION` env constant for Python installations.

This change ensures that the environment in which the coverage
combine/report job runs is executed in the same as the other ones.
This consistency prevents the coverage error:
`No source for code: ...`
which is caused by downloaded coverage data being processed in a
different directory structure than the one it was created in.